+++ FrequencyMagnet/Source/SliderContainer.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "../JuceLibraryCode/JuceHeader.h"
+#include "JuceHeader.h"
 #include "FreqSlider.h"
 #include "WidthSlider.h"
 #include "BiasSlider.h"
@@ -14,13 +14,15 @@ class SliderContainer : public ParameterContainerComponent
 {
 public:
 	SliderContainer(std::shared_ptr<FrequencyMagnetParameters> valueTreeState, Colour textColour, int textBoxHeight);
-    ~SliderContainer();
+    ~SliderContainer() override;
 
     std::shared_ptr<PluginParameters> getPluginParameters() override { return params; }
 	const int getComponentHeight() override { return ParameterContainerComponent::getComponentHeight() * 3; }
     void paint (Graphics&) override;
     void resized() override;
 	Array<PropertyComponent*> getSettingsProperties() override;			
+    void onPropertiesChanged() override;
+    void onAudioValueTreeStateLoadedFromXmlState(PluginParameters* newState, XmlElement* xmlState) override;
 
 private:
 	FreqSlider freqSlider;