+++ PhaseLock/Source/PhaseLockInteractor.cpp
@@ -4,10 +4,31 @@
 #include "PhaseLockFFTProcessor.h"
 
 std::unique_ptr<StandardFFTProcessor> PhaseLockInteractor::createSpectralProcess(int index, int fftSize, int hopSize,
-    int sampleRate, int numOverlaps, int chans, int numChans)
+    int sampleRate, int numOverlaps, int, int)
 {    
-	return std::make_unique<PhaseLockFFTProcessor>(fftSize, hopSize, hopSize * (index%numOverlaps), (int)sampleRate, this->getPhaseBuffer());
+	return std::make_unique<PhaseLockFFTProcessor>(fftSize, hopSize, hopSize * (index%numOverlaps), sampleRate, this->getPhaseBuffer());
 }
+
+void PhaseLockInteractor::process(SpectralAudioPlugin *plugin, std::vector<std::vector<float>> *input, std::vector<std::vector<float>> *output)
+{
+    if (m_params->willContinuePlayingWhenSilentInput())
+    {
+        const auto* playHead = plugin->getPlayHead();
+        if(playHead != nullptr)
+        {
+            auto position = playHead->getPosition();
+            if(position.hasValue()) 
+            {
+                if(!position->getIsPlaying())
+                {
+                    return;
+                }
+            }
+        }
+    }
+    this->SpectralAudioProcessorInteractor::process(input, output);
+}
+
 
 void PhaseLockInteractor::prepareProcess(StandardFFTProcessor * spectralProcessor)
 {