+++ SinusoidalShapedFilter/Source/SSFAudioPlugin.cpp
@@ -5,7 +5,7 @@ public:
     Factory() : m_pluginParams(nullptr) {
     }
     
-    ~Factory(){}
+    ~Factory() override {}
     
     std::shared_ptr<PluginParameters> createParams(SpectralAudioPlugin* plugin) override {
         if(!m_pluginParams) {
@@ -15,12 +15,12 @@ public:
         return m_pluginParams;
     }
 
-    ParameterContainerComponent* createUi(SpectralAudioPlugin* plugin) override {
+    ParameterContainerComponent* createUi(SpectralAudioPlugin*) override {
         return new SliderContainer(m_pluginParams, Colour::fromString(TEXT_COLOUR), 30);
     }
-    std::unique_ptr<SpectralAudioProcessorInteractor> createProcessor(SpectralAudioPlugin* plugin) override {
+    std::unique_ptr<SpectralAudioProcessorInteractor> createProcessor(SpectralAudioPlugin*) override {
         return std::make_unique<SSFInteractor>(SpectralAudioPlugin::FFT_OVERLAPS, m_pluginParams);
-    };
+    }
     
 private:
     std::shared_ptr<SSFParameters> m_pluginParams;
@@ -30,5 +30,12 @@ private:
 // This creates new instances of the plugin..
 AudioProcessor* JUCE_CALLTYPE createPluginFilter()
 {
+    // To debug AUval
+//#if DEBUG
+//    while (!Process::isRunningUnderDebugger())
+//    {
+//        Thread::sleep(250);
+//    }
+//#endif
     return new SpectralAudioPlugin(new Factory());
 }